backend/requirements.txt
# Fashion Production System - Production Requirements
# v5: force cache bust

# Core Django
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Production Server
gunicorn==21.2.0

# Environment
python-dotenv==1.0.0

# File Processing
Pillow==10.1.0
PyMuPDF==1.23.8

# PDF Generation
WeasyPrint==60.1
xhtml2pdf==0.2.11
reportlab==3.6.13

# PDF Parsing
pdfplumber==0.11.5

# Excel Generation
openpyxl==3.1.2

# AI / OpenAI
openai==1.12.0

# Static files
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.39.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
